Add global dependencies injection pattern
This pattern consists of AppDependencies and dependencies protocols. The
AppDependencies is responsible for containing all UILayer dependencies
that point to DataLayer. This is the single point in the app where
dependency flow is violated - which is done on purpose.

Dependencies protocols carry all dependencies that are needed in component
respectively.